feat: 规范类和方法命名
