+++ src/main/resources/io/jenkins/plugins/remote/result/trigger/RemoteBuildResultTriggerScheduledAction/index.jelly
